refactor: governance tools directory
