+++ src/server/server.ts
@@ -19,7 +19,7 @@ import { addCompaniesCapabilities } from '../tools/companies'
 import { addConfigurationCapabilities } from '../tools/configuration'
 import { addDeployCapabilities } from '../tools/deploy'
 import { addMarketplaceCapabilities } from '../tools/marketplace'
-import { addProjectsCapabilities } from '../tools/projects'
+import { addProjectsCapabilities } from '../tools/governance'
 import { addRuntimeCapabilities } from '../tools/runtime'
 import { addServicesCapabilities } from '../tools/services'
 import { APIClient } from '../lib/client'
