--- src/types/project_draft.ts
@@ -1,34 +1,0 @@
-// Copyright Mia srl
-// SPDX-License-Identifier: Apache-2.0
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import { IEnvironment, IProject, ITemplate } from '@mia-platform/console-types'
-
-export interface ProjectDraft {
-  environments?: IEnvironment[]
-  environmentsVariables?: IProject['environmentsVariables']
-  enabledServices?: Record<string, unknown>
-  staticSecret?: ITemplate['staticSecret']
-  templateId?: string
-  repository?: {
-    providerId?: string
-    provider?: {
-      label?: string
-      type?: string
-    },
-    gitPath?: string
-    visibility?: string
-  },
-  pipelines?: Pick<IProject, 'pipelines'>
-}
